delete api completed and data type changes from Long to Integer for id parameter

+++ search/src/main/java/com/project/search/repository/StuffFindRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.project.search.model.StuffFindModel;
 
 @Repository
-public interface StuffFindRepository extends JpaRepository<StuffFindModel, Long> {
+public interface StuffFindRepository extends JpaRepository<StuffFindModel, Integer> {
 	
 	
 	@Query(value = "SELECT * FROM stuff_data  WHERE name LIKE %:name% " ,nativeQuery=true )
